sendofflinemessage
